Allow asynchronous service callbacks
When advertising a service with roslibjs, the service callback returns a
boolean value and fills out a response. Immediately upon returning, this
response is sent back to the service callee. This model, while fine for
quick service callbacks, prevents service callbacks that take time and
must run asynchronously. As a result, a service callback could not (for
example) call another service as part of it's computation before
responding.

This patch allows this, by using the Javascript Promise API. Service
callbacks can now return either a boolean or a promise -- if a boolean,
the service response is sent immediately (as before). If a promise, the
service response is only sent when the promise resolves. This allows for
asynchronous code to be used within service callbacks!