update shim files to use window instead of global
The files in /shim is only used in browsers (remapped in the browser
field of package.json) when using build tools. Browsers do not have the
global propery which is a node thing. Browsers instead have the global
namespace window.

This commit solve my problem of importing roslib in my application and
building with webpack 4 without having webpack shim global = window.

Note: updated Gruntfile.js to have jshint read the config file from the
closes directory/parent directory and added a new .jshintrc to /shim to
lint the files there with a browser environment in mind.