Fix TypeError: this.ros.ActionClient is not a function (#340)
I'm using the ES6 import statement to load roslib which works fine for
most roslib modules, except `TFClient`. That module doesn't properly
import it's dependencies but relies on the properties on the ros object.

I rewrote it a little bit to make it work for ES6 imports. Should be
backwards compatible.