Update Param.js to intuitively request JSON data
If you receive a parameter with get, the data will be received as a number 0, but if you try to modify it with set, you need to JSON.stringify the value again. For example, setting it as {name:"intParam",value:0} will not work, but setting it as {name:"intParam",value:`\"0\"`} will work.