Enable `noUncheckedIndexAccess`
One of the two most important TypeScript rules, imo.
Accessing indexed types (maps, arrays, etc.) in JavaScript
doesn't ever throw an out of bounds error, but it might return
`undefined` - and TypeScript by default doesn't check against
this because so much JavaScript code in the world is written
sloppily enough that checking this by default would make it
hard to port to TypeScript.